requirements.txt
opencv-python==4.5.3.56
opencv-contrib-python==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
scikit-learn==1.0
